+++ tests/TranslationTest.php
@@ -14,6 +14,8 @@ class TranslationTest extends TestCase
      */
     public function it_translates_the_page_word()
     {
+        $this->registerDefaultRoute();
+
         $response = $this->callRoute('/pagina/2');
 
         $this->assertNotEmpty($response['models']['data']);
@@ -24,6 +26,8 @@ class TranslationTest extends TestCase
      */
     public function it_doesnt_accept_the_english_word_anymore_when_translated()
     {
+        $this->registerDefaultRoute();
+        
         $this->setExpectedException('Symfony\Component\HttpKernel\Exception\NotFoundHttpException');
 
         $this->call('GET', 'dummies/page/1')->status();